Move the ecmd to pdbg core mapping logic to common code
